refactor: move AttributeFieldType to core model
Promotes the attribute-type classification enum from the UI settings
package into name.abuchen.portfolio.model so both the desktop UI and the
REST API can share one source of truth. The localized display label moves
to a UI-side AttributeFieldTypeLabels helper.
